requirements.txt
# Core dependencies - Platform-specific backend selection
torch>=2.1.0; sys_platform != 'darwin'
torchvision>=0.16.0; sys_platform != 'darwin'
mlx>=0.28.0; sys_platform == 'darwin'
mlx-nn>=0.1.0; sys_platform == 'darwin'

# Cross-platform dependencies
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
tqdm>=4.65.0
tensorboard>=2.13.0

# 3D and geometry
trimesh>=3.21.0
open3d>=0.17.0

# Scientific computing
scipy>=1.11.0
scikit-image>=0.21.0

# Configuration and logging
omegaconf>=2.3.0
wandb>=0.15.0
hydra-core>=1.3.0

# Development and testing
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0

# Optional: for advanced features
lpips>=0.1.4
kornia>=0.7.0
